requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
numpy
sentence-transformers
requests
